docs: transfer the track constant to a separate module
