+++ roulette/models/seriesModel.js
@@ -7,6 +7,7 @@ let maxBet = localStorage.getItem('maxBet');
 minmax.value = `${minBet}-${maxBet}`;
 // выбор минимума максимума рулетки:
 minmax.onclick = function () {
+    // повторяющийся код:
     // выбор минимума-максимума рулетки:
     let minmax = document.getElementById('minmax').value; // выбираем элемент select minmax
     minBet = localStorage.getItem('minBet');
@@ -42,6 +43,7 @@ minmax.onclick = function () {
             localStorage.setItem('minBet', 1);
             localStorage.setItem('maxBet', 100);
     };
+    //
 };
 // кнопка Сброс:
 reset.onclick = function () {
